requirements.txt
# Remote Whisper Server Requirements

# Flask web framework
flask==3.0.0
flask-cors==4.0.0

# Whisper STT
openai-whisper==20231117
torch==2.5.1
torchaudio==2.5.1

# Audio processing
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
